requirements.txt
asgiref==3.6.0
beautifulsoup4==4.11.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
cryptography==39.0.1
defusedxml==0.7.1
Django==4.1.6
django-allauth==0.52.0
django-bootstrap-v5==1.0.11
django-mathfilters==1.0.0
gunicorn==20.1.0
icalendar==5.0.5
idna==3.4
oauthlib==3.2.2
psycopg2-binary==2.9.5
pycparser==2.21
PyJWT==2.6.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
soupsieve==2.4
sqlparse==0.4.3
urllib3==1.26.14
